@@ -5,8 +5,11 @@
     "start": "webpack-dev-server --hot --progress"
   },
   "dependencies": {
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-query": "^3.39.3",
+    "react-router-dom": "^6.23.1",
+    "recharts": "^2.12.7"
   },
   "devDependencies": {
     "@babel/core": "^7.22.7",
@@ -19,6 +22,7 @@
     "@types/jest": "^29.5.2",
     "babel-jest": "^29.6.0",
     "babel-loader": "^9.1.2",
+    "css-loader": "^7.1.2",
     "html-webpack-plugin": "^5.5.3",
     "jest": "^29.6.0",
     "jest-transform-stub": "^2.0.0",
