@@ -1,7 +1,28 @@
 import React from 'react';
+import { QueryClient, QueryClientProvider } from 'react-query';
 
-const App = () => <h1>UK Energy Mix</h1>;
+// Components
+import InitialPage from './components/initialPage';
+import EnergyPage from './components/energyPage';
 
-export {
-    App
-}
+// Styles
+import './styles.css'
+
+const queryClient = new QueryClient();
+
+const App = () => {
+    return (
+        <QueryClientProvider client={queryClient}>
+            <div>
+                <div className="container-main-title">
+                    <h1 className="main-title">UK Energy Mix</h1>
+                </div>
+                <div className="container-body">
+                    <InitialPage />
+                </div>
+            </div>
+        </QueryClientProvider>
+    );
+};
+
+export default App;
